requirements.txt
python-dotenv==0.21.0
Faker==15.3.4